feat(TurnService): extract move pipeline and turn management
refactor(GameService): delegate move and turn methods to TurnService

test: integrate TurnService into TestServiceFactory
